Allow a port when spawning through the SecureProtocol.
Removed the restriction on commands being SpawnCommand objects.

This requires that the port is set through the uri of the SpawnCommand
as well as the uri for the connection through the use of the
spawn_server method. This reduces the time to connect by getting rid
of the request for the remote server port.

Only three possibilities are allowed:
1) Neither the command uri nor the connection uri have the port set.
 - The port is determined dynamically and output back to the client.
2) The command uri has the port set and the connection uri does not.
 - The port is used from the command uri and is output back to the client.
3) The command uri and the connection uri have the same port set.
 - The port is used statically on both ends without any extra messaging.
*) Any other possibilities will most likely result in runtime errors.