Provide auto-close methods for opening a service, connection, or spawn
Service#spawn_on, when given a block, will shutdown the spawned service
after yielding a proxy to the spawned service to the block.

Service#connect_to, when given a block, will close its proxy automatically
after yielding the created proxy to the block.

Service.open will close the service after yielding itself to a block.

Service.start will wait for the service to close by other means.
This could happen through a trap('INT') that closes the server or
in a manner similar to how spawned services are closed; the front object
has a method that shuts the server down.